Version 0.091
Changed the level size, and the mapping of pond tiles. The art for grass
tiles and pond tiles has been changed to look like what they are. I've
also added the screen buffer to hold the full level size, and made the
program loop around if statements waiting for key presses for movement
of the screen.