Version 0.132
Started changing out the methods that were in the Pixel class for the
methods that are now in the Color Handler class. I've gotten rid of the
unnecessary methods in the main class, and started working through the
level class. I also added a new method to the Pixel class for handling
times at which you want to write a pixel without specifying the position
where it is to be written.