Version 0.142
I finished making the scaling of the tiles dynamic so it can support
multiple widths and heights. I also worked out most of the bugs related
to that. When I changed all these numbers over the attacking animation
wasn't mapping to the right pixels, but I think I've found the solution
that has been implemented in the up arrow animation, while the rest have
yet to be changed.